moved update validator exchange rate
